% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{fun.2plm}
\alias{fun.2plm}
\title{The definition of two parameter log-logistic model function}
\usage{
fun.2plm(alp, x, logarg)
}
\arguments{
\item{alp}{the parameter of 2PLM, 2-dim}

\item{x}{the independent variable}

\item{logarg}{to choose the parameterzation form of 2PLM}
}
\value{
a 2PLM value of specified point
}
\description{
The definition of two parameter log-logistic model(2PLM) function. The log-logistic models are the most used dose-response models.
}
\examples{
\dontrun{
fx<-fun.2plm(c(2,2),seq(0.1, 20, len = 10),1)
} 

}
