% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{fun.3plm}
\alias{fun.3plm}
\title{The definition of three parameter log-logistic model function}
\usage{
fun.3plm(alp, x, logarg)
}
\arguments{
\item{alp}{the parameter of 3PLM, 3-dim}

\item{x}{the independent variable}

\item{logarg}{to choose the parameterzation form of 3PLM}
}
\value{
a 3PLM value of specified point
}
\description{
The definition of three parameter log-logistic model(3PLM) function. The log-logistic models are the most used dose-response models.
}
\examples{
\dontrun{
fx<-fun.3plm(c(2,2,5),seq(0.1, 20, len = 10))
} 
}
