% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{fun.5plm}
\alias{fun.5plm}
\title{The definition of five parameter log-logistic model(5PLM) function.}
\usage{
fun.5plm(alp, x, logarg = 1)
}
\arguments{
\item{alp}{the parameter of 5PLM, 5-dim}

\item{x}{the independent variable}

\item{logarg}{to choose the parameterzation form}
}
\description{
The definition of five parameter log-logistic model(5PLM) function.
}
