% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, thin, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{N}{the number of samples (int)}

\item{thin}{the number of between-sample random numbers (int)}

\item{mu1}{the mean of first marginal value (double)}

\item{mu2}{the mean of second marginal value (double)}

\item{sigma1}{the variance of first marginal value (double)}

\item{sigma2}{the variance of second marginal value (double)}

\item{rho}{correlation (double)}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <-gibbsC(5000,2,0,0,1,1,.9)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
