% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{Gibbsinr}
\alias{Gibbsinr}
\title{A Gibbs sampler using R}
\usage{
Gibbsinr(N, burn, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{N}{the number of samples}

\item{burn}{burn-in length, the length of deserted samples from the first}

\item{mu1}{the mean of first marginal value}

\item{mu2}{the mean of second marginal value}

\item{sigma1}{the variance of first marginal value}

\item{sigma2}{the variance of second marginal value}

\item{rho}{correlation}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
rnR <- <-Gibbsinr(5000,0,0,0,1,1,.9)
par(mfrow=c(2,1));
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
