% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{obj.gr.3pl}
\alias{obj.gr.3pl}
\title{3PLM's gradient}
\usage{
obj.gr.3pl(p, x)
}
\arguments{
\item{p}{the parameter of 3PLM, 3-dim}

\item{x}{the independent variable}
}
\description{
Calculate 3PLM's gradient of specified point
}
\examples{
\dontrun{
3plm.grad<-obj.gr.3pl(c(2,2,5),seq(0.1, 20, len = 10))
} 
}
