% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{obj.hess.2pl}
\alias{obj.hess.2pl}
\title{2PLM's Hessian matrix}
\usage{
obj.hess.2pl(p, x)
}
\arguments{
\item{p}{the parameter of 2PLM, 2-dim}

\item{x}{the independent variable}
}
\description{
Calculate 2PLM's Hessian matrix of specified point
}
