% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{obj.hess.3pl}
\alias{obj.hess.3pl}
\title{3PLM's Hessian matrix}
\usage{
obj.hess.3pl(p, x)
}
\arguments{
\item{p}{the parameter of 3PLM, 3-dim}

\item{x}{the independent variable}
}
\description{
Calculate 3PLM's Hessian matrix of specified point
}
