% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{fun.3plm}
\alias{fun.3plm}
\title{The definition of three parameter log-logistic model(3PLM) function.}
\usage{
fun.3plm(alp, x, logarg = 1)
}
\arguments{
\item{alp}{the parameter of 3PLM, 3-dim}

\item{x}{the independent variable}

\item{logarg}{to choose the parameterzation form}
}
\description{
The definition of three parameter log-logistic model(3PLM) function.
}
