% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{fun.4plm}
\alias{fun.4plm}
\title{The definition of four parameter log-logistic model(4PLM) function.}
\usage{
fun.4plm(alp, x, logarg = 1)
}
\arguments{
\item{alp}{the parameter of 4PLM, 4-dim}

\item{x}{the independent variable}

\item{logarg}{to choose the parameterzation form}
}
\description{
The definition of four parameter log-logistic model(4PLM) function.
}
