% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{fun.5plm}
\alias{fun.5plm}
\title{The definition of five parameter log-logistic model function}
\usage{
fun.5plm(alp, x, logarg)
}
\arguments{
\item{alp}{the parameter of 5PLM, 5-dim}

\item{x}{the independent variable}

\item{logarg}{to choose the parameterzation form of 5PLM}
}
\value{
a 5PLM value of specified point
}
\description{
The definition of five parameter log-logistic model(5PLM) function. The log-logistic models are the most used dose-response models.
}
\examples{
\dontrun{
fx<-fun.5plm(c(2,2,5,1,3),seq(0.1, 20, len = 10))
} 
}
