% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{obj.gr.5pl}
\alias{obj.gr.5pl}
\title{5PLM's gradient}
\usage{
obj.gr.5pl(p, x)
}
\arguments{
\item{p}{the parameter of 5PLM, 5-dim}

\item{x}{the independent variable}
}
\description{
Calculate 5PLM's gradient of specified point
}
\examples{
\dontrun{
5plm.grad<-obj.gr.5pl(c(2,2,5,1,3),seq(0.1, 20, len = 10))
}
}
