% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.r
\name{obj.hess.5pl}
\alias{obj.hess.5pl}
\title{5PLM's Hessian matrix}
\usage{
obj.hess.5pl(p, x)
}
\arguments{
\item{p}{the parameter of 5PLM, 5-dim}

\item{x}{the independent variable}
}
\description{
Calculate 5PLM's Hessian matrix of specified point
}
